Fixed unmanaged chains
My understanding of the `iptables-restore --noflush` command was flawed.
It turns out that `iptables-restore` will flush the chain if it already
exists, and there appears to be no way to simply ensure that a chain
exists for non-builtin chains.

So we fix that by introducing a generator that only outputs a list of
`(table, chain)` pairs to be created, and another generator that will
output a ruleset with no chain definitions that can then be applied via
`iptables-restore -n`.